# Core: Boot Vitest runner when hovering the Testing Module